added integer data type

+++ plugins/wasi_nn/wasinn_ggml.h
@@ -85,7 +85,7 @@ struct Graph {
   std::string Grammar;
   uint64_t Seed = LLAMA_DEFAULT_SEED;
   // Configs.
-  double TopK;
+  int32_t TopK;
   LocalConfig Conf;
 };
